✨ Feat: 모임 기록 활동 장소 연결
Related to: #315

+++ app/src/main/java/com/mongmong/namo/data/dto/ActivityDTO.kt
@@ -25,9 +25,9 @@ data class ActivityParticipant(
 
 data class ActivityLocation(
     val kakaoLocationId: String,
-    val latitude: Int,
+    val latitude: Double,
     val locationName: String,
-    val longitude: Int
+    val longitude: Double
 )
 
 data class ActivityImage(
